GoPiGo/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
groq>=0.9.0
python-multipart>=0.0.9
opencv-python>=4.9.0.0
numpy>=1.26.0
pydantic>=2.6.0
python-dotenv>=1.0.0
